+++ udata/tests/api/test_datasets_api.py
@@ -1250,6 +1250,28 @@ class DatasetAPITest(APITestCase):
             "version": None,
         }
 
+    def test_remove_schema(self):
+        self.login(AdminFactory())
+        resource = ResourceFactory(schema={"name": "etalab/schema-irve-statique"})
+        dataset = DatasetFactory(resources=[resource])
+
+        # not sending the schema should keep the current one
+        data = {"title": "updated 1"}
+        response = self.put(url_for("api.resource", dataset=dataset, rid=str(resource.id)), data)
+        self.assert200(response)
+        dataset.reload()
+        assert dataset.resources[0].title == "updated 1"
+        assert dataset.resources[0].schema is not None
+
+        # sending a None schema should remove it
+        data = {"title": "updated 2", "schema": None}
+        response = self.put(url_for("api.resource", dataset=dataset, rid=str(resource.id)), data)
+        self.assert200(response)
+
+        dataset.reload()
+        assert dataset.resources[0].title == "updated 2"
+        assert dataset.resources[0].schema is None
+
 
 class DatasetsFeedAPItest(APITestCase):
     def test_recent_feed(self):